Initial Commit of React Redux to get /r/aww
Simple app to pull from /r/aww, and render using react-redux.

I want to display the first X posts in /r/aww, filtering out NSFW
posts and non-image posts.

On load, I want the first image to be displayed as large as
possible, and the image list to be clickable. When clicked,
the selected image will become the main image.